fix: few minor redux refs and cleanup
