+++ src/components/games/blackjack/Blackjack.tsx
@@ -7,7 +7,13 @@
 import React from "react";
 import { useRecoilState } from "recoil";
 import asyncForEach from "../../../helpers/asyncForEach";
-import { weighHand, playBot, banking, DEALER } from "./blackjackHelpers";
+import {
+  weighHand,
+  playBot,
+  banking,
+  DEALER,
+  getGameFunctions,
+} from "./blackjackHelpers";
 import Header from "./Header";
 import GameTable from "../game-table";
 import Deck from "../../../apis/Deck";
@@ -26,42 +32,6 @@ const BlackJack: React.FC = () => {
     },
     setState,
   ] = useRecoilState(blackjackState);
-
-  /** get the game functions for the present hand */
-  const getGameFunctions = (hand: DBHand): GameFunctions[] => {
-    if (!hand) {
-      return [];
-    }
-
-    // reset game functions
-    const newGameFunctions = [GameFunctions.STAY];
-    const handWeight = hand.weight || 0;
-
-    // check if not a bust
-    if (handWeight < 21) {
-      newGameFunctions.push(GameFunctions.HIT);
-      // check if you only have 2 cards
-      if (hand.cards.length === 2) {
-        newGameFunctions.push(GameFunctions.DOUBLE);
-        // check if card1 and card2 have equal weight
-        const { weight: weight1 } = weighHand([hand.cards[0]]);
-        const { weight: weight2 } = weighHand([hand.cards[1]]);
-        if (weight1 === weight2) {
-          newGameFunctions.push(GameFunctions.SPLIT);
-        }
-      }
-    }
-
-    return newGameFunctions;
-
-    // update game state
-    // setState({
-    //   bj: {
-    //     gameFunctions: newGameFunctions,
-    //     hideHands,
-    //   },
-    // });
-  };
 
   /** function that takes a hand of duplicates and makes 2 hands */
   const split = async (): Promise<void> => {
